Move BubbleWrap::HTTP::Query and BubbleWrap::HTTP::Response into individual files under http/
