+++ lib/bubble-wrap/http.rb
@@ -2,3 +2,5 @@ require 'bubble-wrap/loader'
 BubbleWrap.require('motion/core/ns_url_request.rb')
 BubbleWrap.require('motion/core.rb')
 BubbleWrap.require('motion/http.rb')
+BubbleWrap.require('motion/http/query.rb')
+BubbleWrap.require('motion/http/response.rb')
